flytt skjemanummer-koden til en egen fil, for bedre lesbarhet
